Replace package log-config.edn template with log4j configuration
Issue #510 - Remove log-config.edn.template from the package and
add a log4j2.xml file to configure log4j. Update drafter-start.sh
to specify this file as the configuration file to use rather than
the log4j2.xml file packaged in the drafter JAR.

The log4j2.xml file allows the location of the logs directory to
be specified via the optional DRAFTER_LOGS_DIR environment variable.
Add a drafter/logs-dir key to the package config for this variable
and remove the old keys for specifying the location of the
drafter.log and sparql.log files directly.

Update the logging documentation to describe configuring logging in
production.