Define enum option for gzip option
Issue #537 - Define an enum for the gzip option to client/add-data.
This allows the uesr to request the input be compressed when added
to to the request or state that it is already in a compressed
format.

Attempt to infer the properties of the input if the :format and
:gzip options are not specified. If the input is a file, check
for a .gz or .gzip extension to indicate if the input is already
compressed. If so, remove it and use the remaining extension to
infer the RDF format. Combine the inferred properties with the
user-specified options to decide how to handle compressing the input
data and setting the Content-Type and Content-Encoding headers
on the request.