Use our stardog container with the license file layered in for CI
NOTE: this container will be vulnerable to log4shell, unlike our
patched omni builds - so do not use this in prod.

Remove steps to start stardog now we have it as a container service

Prior to this change I did try some experiments consuming complexibles
stardog docker image from their artifactory instance with the
co-ordinate:

stardog-eps-docker.jfrog.io/stardog:6.2.3

However there is no way with circleci's docker executor affordances to
call docker exec on the image to build the databases, or to include
the license in a volume.

So instead we use our own customised stardog docker image (built by
hand) as part of this PR https://github.com/Swirrl/stardog/pull/25

This image exposes a `CREATE_DATABASE` environment variable which if
set to a database name will construct that database before starting
the image, allowing it to be used with CircleCI docker executors.