Configure same db name everywhere in CI
This is because our hack to the swirrl/stardog container only takes
one db name as an argument. However we can get away with it in CI
because we have parallel jobs, so db's are isolated into different
jobs, and can therefore use the same db name.

We now make the test db configurable via env (previously it was
hardcoded in some places)