Rewrite multiple graph references within draftset statements
Issue #607 - Change how draftset query rewriting works to handle
statements which contain multiple graph references to graphs within
the draftset.

The existing query searches for statements which contain a single
reference to a graph within the draft in either the subject,
predicate or object position. This query does not work as expected
for statements which contain two or three components which reference
a graph within the draftset. For these statements, multiple UNION
clauses match the source statement and each component is rewritten
independently, which leads to two or three partially-rewritten
statements within the draft. Running the rewriting query multiple
times causes these extraneous statements to be gradually rewritten
towards their fully-rewritten state, so these may not always be
visible when inserting or deleting data.

Split the current query into three separate queries which rewrite
all subjects, predicates and objects within each statement in the
draftset independently. This prevents the multiple partial match
behaviour of the old query, and also works for statements where
each component references a diffferent graph e.g.

    GRAPH <http://g1> {
        <http://s> <http://g1> <http://g2> .
    }

Add low-level tests for the behaviour of the rewrite-draftset! and
unrewrite-draftset! functions used to rewrite a draft after an
update.

Add higher-level tests which add statements referencing draftset
graphs within a draftset and checks they are published correctly.