Configure swagger security definitions with authentication methods
Issue #581 - Add methods to the AuthenticationMethod protocol to
configure the swagger security definitions and aspects of the
swagger UI. Each configured authentication method is identified by
a key and constructs a security scheme object which is injected into
the base swagger configuration defined in drafter.yaml. Each
authentication method also constructs a security requirement object
for each API operation.

The requried security definitions for each operation depend on
whether global authentication is configured. If it is, all routes
require authentication, otherwise some routes such as the public
endpoint and SPARQL query endpoints are unauthenticated. Add a
short definition of each API operation in the drafter.swagger
namespace and use it to configure the security requirements for
each authentication method.

Remove the exising security method descriptions from drafter.yml.
Add markdown descriptions for each authentication method explaining
how to authenticate HTTP requests and via the UI. Move the
'Authentication and Security' section of the swagger description to
the end and append the descriptions for each configured
authentication method.

Add a ring-swagger dependency and use it to create the routes for
hosting the swagger UI. Add a drafter.swagger/swagger-routes
integrant component for definining all the swagger routes. Inject
this component as a dependency of the main drafter.handler/app
handler component.

Add custom swagger UI index.html resource to be used instead of the
one in ring-swagger-ui. This page installs a custom request
interceptor which adds the audience request parameter when calling
the auth0 token endpoint. The audience is a parameter of the auth0
authentication methods, and they inject this into the swagger UI
configuration via the get-swagger-ui-config method which enables
each authentication method to configuret the behaviour of the UI.

Add a 'mock auth0' authentication method which uses the shared RSA
keypair to generate and validate JWT tokens. Token parsing and
validation work in the same way as the auth0 authentication method,
but within the swagger UI the mock method is presented as an API key
method instead of OAuth.