Log SPARQL queries to a separate log file
Issue #615 - Log messages within the following namespaces to a
separate sparql.log file:

  * drafter.rdf.sparql-protocol
  * drafter.rdf.sparql
  * drafter.feature.draftset.update
  * drafter.routes.sparql
  * drafter.backend.draftset.rewrite-query
  * drafter.stasher

The sparql.log file is approximately 2-3 times the size of
drafter.log so it is configured to store a larger number of backup
files in the default log configuration.

Add log messages to drafter.feature.draftset.update to log the raw
and re-written UPDATE queries for consistency with the
sparql-protocol namespace used for querying.

Implement Object.toString for re-written SPARQL query records so
these queries are shown in the logs.