Add tools.build tasks to build jar, pack and uberjar
Issue #635 - Add 'jar' and 'install' tasks to build.clj. The jar
tasks builds a jar for the current drafter version and the install
tasks builds and installs the jar into the local maven repository.

Add an 'uber' task for building an uberjar for the current version.

Add a 'clean' task to clean the target directory.

Add 'skinny' task which builds a skinny jar with pack. Add a :skinny
alias into deps.edn to include the required resources. Update the
pack script to use this task instead of calling pack directly.

The jar, install and uber tasks all support a :type option
indicating which drafter version to build for. This should name
an alias within deps.edn containing the production configuration
to use. This defaults to :pmd4 if not specified.

Exclude servlet-api from lib-noir dependency since this is older
than the one required by jetty. Add top-level dependency for
jackson-core.