Add generate for benchmark test data
Add generator for nquad files containing the specified number of
statements and graphs. The generated statements are distributed
approximately evenly between the requested graphs. Around 10% of
the generated subjects, predicates and objects reference the
graphs in the data.

Add Makefile for generating data files with different number of
statements and graphs. The files contain 1k, 10k, 100k and 1 million
statements split between 1, 10, 100 or 1000 graphs. Each data file
also has a corresponding 'deletion' file containing half of the
input statements to be used by the deletion benchmark.