Set default parameters for each benchmark
Add annotations to each benchmark class to define the default
benchmark settings. Remove the warmup iterations and reduce the
number of forks to 2, and the number of iterations per-fork to 2.
Some of the larger tests take a long time (~1 hour) to run so trying
to run 5 iterations across 5 forks (the default) is much too
time-consuming.