Generate all graphs from benchmark result files
Update the chart generation method to generate all charts for a
collection of result files. Each benchmark is expected to have three
parameters - the total number of statement, the total number of
graphs and the percentage of 'graph-referencing' statements in the
input file.

A chart can be generating by choosing one of these dimensions for
the x-axis and fixing the other two dimension values to find a
collection of benchmark times. If there are multiple values for
a fixed set of dimensions, a chart can be plotted.

Add benchmark-set-slices function which collects all chart data
by fixing each of the statements, graphs and ref-statements
dimensions. This returns a map from keys describing each chart and
the corresponding set of results across all benchmark files.

Add write-chart function which generates the chart heading, axis
labels and file names from the chart key and writes the chart image
to the output directory.