Add generator for benchmark test data
Issue #635 - Add data-gen project for generating test data to be
used by the benchmarks. This exposes a simple CLI interface for
generating a single data file, or all files required by the
benchmarks. Each data file is parameterised by the total number
of statements, the number of graphs and the number or percentage
of 'graph-referencing' statements i.e. statements which reference
a dataset graph in the subject, predicate or object.