Add benchmark harness for the main drafter operations
Issue #635 - Add drafter-benchmarks project for benchmarking the
append, delete, delete graph, publish and SPARQL update operations
within a draftset.

The benchmarks are defined within a java project which uses JMH to
generate benchmark code from annotated benchmark classes and build
an uberjar to run selected benchmarks. The project requires a
dependency on a jar for the drafter version to benchmark. These
can be installed into the local repository using the tools.build
install task within the drafter directory.

Benchmark classes attempt to benchmark the corresponding operation
with as little overhead as possible by calling directly into the
clojure code using the java interop interface. For example the
append and delete benchmarks construct and execute the append or
delete state machines directly and execute these synchronously
rather than go via the worker queue as in the normal drafter
operation. There is a small java wrapper layer around the clojure
draftset and user objects to simplify keeping track of otherwise
untyped objects.

Most benchmark classes define custom JMH state classes which
define setup and teardown operations. Some benchmarks have an
element of randomness - for example the SPARQL update benchmark
constructs a random UPDATE statement to execute. Construction of
the query takes place in the setup stage so as not to affect
the benchmark itself.