Add generator for charts of benchmark results
Issue #635 - Add perf-charts project which generates comparison
charts for a set of drafter benchmark results. Each chart is
generated by fixing two of the benchmark parameters (number of
statements, graphs or percentage of graph-referencing statements)
and visualising the effect of the third parameter on performance.
All non-degenerate charts (i.e. those with more than one data point)
are written to the output directory specified on the command-line.