Create JAR files for dependent lib directories in skinny task
Issue #640 - Skinny jars are usually run with a classpath of
'lib/*:drafter.jar' i.e. the main drafter JAR and all dependent
JARs within the lib directory.

The swirrl auth0 dependency is included as a git reference and is
not packaged as a jar. This causes it to be written to the lib
directory as a directory by default. This directory is not iteself
included on the classpath. Writing it and the auth0 resource
directories as JAR files result in them being included on the
classpath.